[WIP][DBInstance] Downgrade RdsClient for instances with security groups
