+++ aws-rds-dbinstance/src/main/java/software/amazon/rds/dbinstance/TranslatorV19.java
@@ -3,15 +3,12 @@ package software.amazon.rds.dbinstance;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import org.apache.commons.lang3.BooleanUtils;
 
@@ -24,11 +21,8 @@ import software.amazon.awssdk.services.rds.model.CreateDbInstanceRequest;
 import software.amazon.awssdk.services.rds.model.DBParameterGroupStatus;
 import software.amazon.awssdk.services.rds.model.DBSubnetGroup;
 import software.amazon.awssdk.services.rds.model.DeleteDbInstanceRequest;
-import software.amazon.awssdk.services.rds.model.DescribeDbClustersRequest;
 import software.amazon.awssdk.services.rds.model.DescribeDbEngineVersionsRequest;
-import software.amazon.awssdk.services.rds.model.DescribeDbInstancesRequest;
 import software.amazon.awssdk.services.rds.model.DescribeDbParameterGroupsRequest;
-import software.amazon.awssdk.services.rds.model.DescribeDbSnapshotsRequest;
 import software.amazon.awssdk.services.rds.model.ModifyDbInstanceRequest;
 import software.amazon.awssdk.services.rds.model.RebootDbInstanceRequest;
 import software.amazon.awssdk.services.rds.model.RemoveRoleFromDbInstanceRequest;
@@ -36,32 +30,7 @@ import software.amazon.awssdk.services.rds.model.RestoreDbInstanceFromDbSnapshot
 import software.amazon.awssdk.utils.StringUtils;
 import software.amazon.rds.common.handler.Tagging;
 
-public class Translator {
-
-    public static DescribeDbInstancesRequest describeDbInstancesRequest(final ResourceModel model) {
-        return DescribeDbInstancesRequest.builder()
-                .dbInstanceIdentifier(model.getDBInstanceIdentifier())
-                .build();
-    }
-
-    public static DescribeDbClustersRequest describeDbClustersRequest(final ResourceModel model) {
-        return DescribeDbClustersRequest.builder()
-                .dbClusterIdentifier(model.getDBClusterIdentifier())
-                .build();
-    }
-
-    public static DescribeDbInstancesRequest describeDbInstancesRequest(final String nextToken) {
-        return DescribeDbInstancesRequest.builder()
-                .marker(nextToken)
-                .build();
-    }
-
-    public static DescribeDbSnapshotsRequest describeDbSnapshotsRequest(final ResourceModel model) {
-        return DescribeDbSnapshotsRequest.builder()
-                .dbSnapshotIdentifier(model.getDBSnapshotIdentifier())
-                .build();
-    }
-
+public class TranslatorV19 extends BaseTranslator {
 
     public static CreateDbInstanceReadReplicaRequest createDbInstanceReadReplicaRequest(
             final ResourceModel model,
@@ -135,14 +104,6 @@ public class Translator {
                 .build();
     }
 
-    public static Integer getAllocatedStorage(final ResourceModel model) {
-        Integer allocatedStorage = null;
-        if (StringUtils.isNotBlank(model.getAllocatedStorage())) {
-            allocatedStorage = Integer.parseInt(model.getAllocatedStorage());
-        }
-        return allocatedStorage;
-    }
-
     public static CreateDbInstanceRequest createDbInstanceRequest(
             final ResourceModel model,
             final Tagging.TagSet tagSet
@@ -202,7 +163,7 @@ public class Translator {
             final ResourceModel desiredModel,
             final Boolean isRollback
     ) {
-        ModifyDbInstanceRequest.Builder builder = ModifyDbInstanceRequest.builder()
+        final ModifyDbInstanceRequest.Builder builder = ModifyDbInstanceRequest.builder()
                 .allowMajorVersionUpgrade(desiredModel.getAllowMajorVersionUpgrade())
                 .applyImmediately(Boolean.TRUE)
                 .autoMinorVersionUpgrade(desiredModel.getAutoMinorVersionUpgrade())
@@ -350,7 +311,7 @@ public class Translator {
             final List<software.amazon.awssdk.services.rds.model.DBInstance> dbInstances
     ) {
         return streamOfOrEmpty(dbInstances)
-                .map(Translator::translateDbInstanceFromSdk)
+                .map(TranslatorV19::translateDbInstanceFromSdk)
                 .collect(Collectors.toList());
     }
 
@@ -360,7 +321,7 @@ public class Translator {
         final String dbParameterGroupName = Optional.ofNullable(dbInstance.dbParameterGroups())
                 .orElse(Collections.emptyList())
                 .stream()
-                .map(Translator::translateDBParameterGroupFromSdk)
+                .map(TranslatorV19::translateDBParameterGroupFromSdk)
                 .findFirst().orElse(null);
 
         // {@code DbInstance.port} can contain a null-value, in this case we
@@ -462,32 +423,6 @@ public class Translator {
                 .collect(Collectors.toList());
     }
 
-    public static List<Tag> translateTagsFromSdk(final Collection<software.amazon.awssdk.services.rds.model.Tag> sdkTags) {
-        return streamOfOrEmpty(sdkTags)
-                .map(tag -> Tag
-                        .builder()
-                        .key(tag.key())
-                        .value(tag.value())
-                        .build()
-                )
-                .collect(Collectors.toList());
-    }
-
-    public static Map<String, String> translateTagsToRequest(final Collection<Tag> tags) {
-        return streamOfOrEmpty(tags)
-                .collect(Collectors.toMap(Tag::getKey, Tag::getValue, (v1, v2) -> v2, LinkedHashMap::new));
-    }
-
-    public static Set<software.amazon.awssdk.services.rds.model.Tag> translateTagsToSdk(final Collection<Tag> tags) {
-        return streamOfOrEmpty(tags)
-                .map(tag -> software.amazon.awssdk.services.rds.model.Tag.builder()
-                        .key(tag.getKey())
-                        .value(tag.getValue())
-                        .build()
-                )
-                .collect(Collectors.toCollection(LinkedHashSet::new));
-    }
-
     public static List<ProcessorFeature> translateProcessorFeaturesFromSdk(
             final Collection<software.amazon.awssdk.services.rds.model.ProcessorFeature> sdkProcessorFeatures
     ) {
@@ -547,45 +482,6 @@ public class Translator {
                         .roleArn(role.getRoleArn())
                         .build())
                 .collect(Collectors.toList());
-    }
-
-    public static Integer translatePortToSdk(final String portStr) {
-        if (StringUtils.isEmpty(portStr)) {
-            return null;
-        }
-        //NumberFormatException
-        return Integer.parseInt(portStr, 10);
-    }
-
-    public static String translatePortFromSdk(final Integer port) {
-        if (port == null) {
-            return null;
-        }
-        return port.toString();
-    }
-
-    private static <T> Stream<T> streamOfOrEmpty(final Collection<T> collection) {
-        return Optional.ofNullable(collection)
-                .map(Collection::stream)
-                .orElseGet(Stream::empty);
-    }
-
-    private static boolean canUpdateAllocatedStorage(final String fromAllocatedStorage, final String toAllocatedStorage) {
-        if (fromAllocatedStorage == null || toAllocatedStorage == null) {
-            return true;
-        }
-        final int from, to;
-        try {
-            from = Integer.parseInt(fromAllocatedStorage);
-            to = Integer.parseInt(toAllocatedStorage);
-        } catch (NumberFormatException e) {
-            return true;
-        }
-        return to >= from;
-    }
-
-    private static boolean canUpdateIops(final Integer fromIops, final Integer toIops) {
-        return fromIops == null || toIops == null || toIops >= fromIops;
     }
 
     private static boolean shouldSetProcessorFeatures(final ResourceModel previousModel, final ResourceModel desiredModel) {
